intro_database/project/requirements.txt
psycopg2==2.6.1
python-dotenv